# app/static/swagger.yaml
swagger: '2.0'
info: 
  title: "Mechanic Shop API"
  description: "Management from the shop's perspective"
  version: "0.0.1"
host: "127.0.0.1:5000" #during development my API i on local host, will change in prod
schemes: 
  - "http"
consumes: 
  - "application/json"
produces: 
  - "application/json"
securityDefinitions: 
  bearerAuth: 
    type: apiKey
    name: Authorization
    in: header
paths: #information about routes
  /mechanics/login:
    post: 
      tags: 
        - Mechanics
      summary: "Login endpoint (generates token)"
      description: "Validates mechanic credentials, produces a mechanic specific token"
      parameters: #don't forget the s in parameters
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema: 
            $ref: "#/definitions/LoginCredentials"
      responses: 
        200: 
          description: "Login Successful"
          schema: 
            $ref: "#/definitions/LoginResponse"
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Create a new mechanic"
      description: "Creates a mechanic account (requires token)."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/Mechanic"
     
    get:
      tags:
        - Mechanics
      summary: "Read all mechanics"
      description: "Retrieve details for all mechanics"
      responses:
        200:
          description: "List of Mechanic details"
          schema:
            type: array
            items:
              $ref: "#/definitions/Mechanic"

    delete:
      tags:
        - Mechanics
      summary: "Delete the mechanic (token required)"
      description: "Deletes the mechanic linked to the provided token."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Mechanic deleted successfully"
          schema:
            type: object
            properties:
              message:
                type: string
  /mechanics/{mechanic_id}:
    get:
      tags:
        - Mechanics
      summary: "Get mechanic by ID"
      description: "Retrieve a specific mechanic based on ID passed in"
      parameters:
        - in: "path"
          name: "mechanic_id"
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic details"
          schema:
            $ref: "#/definitions/Mechanic"

    put:
      tags:
        - Mechanics
      summary: "Update mechanic"
      description: "Update a mechanic’s details (requires token)."
      security:
        - bearerAuth: []
      parameters:
        - name: "mechanic_id"
          in: "path"
          required: true
          type: integer
        - in: "body"
          name: "body"
          description: "all data for mechanic including updates"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/Mechanic"
...
definitions: #information about data received/produced
  LoginCredentials: #input data required for mechanic login route
    type: object
    properties: 
      email: 
        type: string
      password: 
        type: string

    required: 
      - email
      - password
  LoginResponse: #Response from /login enpdpoint
    type: object
    properties: 
      message: 
        type: string
      token:
        type: string
  MechanicInput:
    type: object
    required:
      - first_name
      - last_name
      - email
      - password
      - salary
      - address
    properties:
      first_name:
        type: string
        example: "John"
      last_name:
        type: string
        example: "Doe"
      email:
        type: string
        example: "john@example.com"
      password:
        type: string
        example: "password123"
      salary:
        type: number
        format: float
        example: 55000.00
      address:
        type: string
        example: "123 Main St, Las Vegas, NV"
  Mechanic:
    type: object
    properties:
      id:
        type: integer
        example: 1
      first_name:
        type: string
        example: "John"
      last_name:
        type: string
        example: "Doe"
      email:
        type: string
        example: "john@example.com"
      salary:
        type: number
        format: float
        example: 55000.00
      address:
        type: string
        example: "123 Main St, Las Vegas, NV"
...